Upgrading to double-time based server time Fixing some warnings

+++ Assets/Core/Scripts/Generic/CyclicAnimatedObject.cs
@@ -9,6 +9,6 @@ public class CyclicAnimatedObject : MonoBehaviour
 
     void Update()
     {
-        animationToCycle.SampleAnimation(gameObject, GameTicker.singleton.predictedServerTime % animationToCycle.length);
+        animationToCycle.SampleAnimation(gameObject, (float)(GameTicker.singleton.predictedServerTime % animationToCycle.length));
     }
 }
